requirements.txt
appdirs==1.4.4
apturl==0.5.2
asteval==1.0.2
attrs==21.2.0
bcrypt==3.2.0
beautifulsoup4==4.10.0
beniget==0.4.1
blinker==1.4
blosc2==2.7.1
Brlapi==0.8.3
Brotli==1.0.9
certifi==2020.6.20
chardet==4.0.0
click==8.0.3
colorama==0.4.4
command-not-found==0.3
contourpy==1.2.1
cryptography==3.4.8
cupshelpers==1.0
cycler==0.11.0
dbus-python==1.2.18
decorator==4.4.2
defer==1.0.6
dill==0.3.8
distro==1.7.0
distro-info==1.1+ubuntu0.2
duplicity==0.8.21
fasteners==0.14.1
fonttools==4.29.1
fs==2.4.12
future==0.18.2
gast==0.5.2
h5py==3.6.0
h5py.-debian-h5py-serial==3.6.0
hdf-compass==0.7b8
html5lib==1.1
httplib2==0.20.2
idna==3.3
importlib-metadata==4.6.4
itk==5.4.0
itk-core==5.4.0
itk-filtering==5.4.0
itk-io==5.4.0
itk-numerics==5.4.0
itk-registration==5.4.0
itk-segmentation==5.4.0
jeepney==0.7.1
keyring==23.5.0
kiwisolver==1.3.2
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lmfit==1.3.2
lockfile==0.12.2
louis==3.20.0
lxml==4.8.0
lz4==3.1.3+dfsg
macaroonbakery==1.3.1
Mako==1.1.3
Markdown==3.6
MarkupSafe==2.0.1
matplotlib==3.5.0
monotonic==1.6
more-itertools==8.10.0
mpmath==0.0.0
msgpack==1.0.8
ndindex==1.8
netifaces==0.11.0
nose==1.3.7
numexpr==2.10.1
numpy==1.24.4
oauthlib==3.2.0
olefile==0.46
packaging==21.3
pandas==2.2.2
paramiko==2.9.3
pexpect==4.8.0
Pillow==9.0.1
ply==3.11
protobuf==3.12.4
psutil==6.0.0
ptyprocess==0.7.0
py-cpuinfo==9.0.0
pycairo==1.20.1
pycups==2.0.1
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
PyQt5==5.15.11
PyQt5-Qt5==5.15.14
PyQt5_sip==12.15.0
pyRFC3339==1.1
python-apt==2.4.0+ubuntu3
python-dateutil==2.9.0.post0
python-debian==0.1.43+ubuntu1.1
pythran==0.10.0
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
QtPy==2.4.1
qtwidgets==1.1
reportlab==3.6.8
requests==2.25.1
scipy==1.8.0
seaborn==0.13.2
SecretStorage==3.3.1
setuptools-scm==8.1.0
six==1.16.0
soupsieve==2.3.1
sympy==1.9
systemd-python==234
tables==3.9.2
tales==0.0.4
tomli==2.0.1
typing_extensions==4.12.2
tzdata==2024.1
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufoLib2==0.13.1
ufw==0.36.1
unattended-upgrades==0.1
uncertainties==3.2.2
unicodedata2==14.0.0
urllib3==1.26.5
usb-creator==0.3.7
vtk==9.3.1
wadllib==1.3.6
webencodings==0.5.1
wxPython==4.0.7
xdg==5
xkit==0.0.0
zipp==1.0.0